% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{assignCellCycle}
\alias{assignCellCycle}
\title{Assign cell cycle}
\usage{
assignCellCycle(
  sce,
  recenter_method = "optim",
  species = "human",
  gname.type = "SYMBOL",
  gname = NULL,
  exprs_values = "logcounts"
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{recenter_method}{character, one of  \code{"optim"} (default),  \code{"chull"}, or  \code{"none"}}

\item{species}{character, one of  \code{"human"} (default) or  \code{"mouse"}.
See \code{\link[tricycle](project_cycle_space)} for more information.}

\item{gname.type}{character, one of  \code{"SYMBOL"} (default) or \code{"ENSEMBL"}.
Will be used to look up the genes in the  \code{rownames(sce)}.
See \code{\link[tricycle](project_cycle_space)} for more information.}

\item{gname}{character, alternative \code{rowData(sce)} column with gene identifiers.
Default is NULL. 
see \code{\link[tricycle](project_cycle_space)} for more information.}

\item{exprs_values}{character, name of the assay in \code{sce} to be used for
projection. Default is \code{"logcounts"}.}
}
\value{
a \code{SingleCellExperiment} object with the \code{tricycleEmbedding} 
   slot in \code{reducedDim(sce)} and \code{tricyclePosition} and 
   \code{tricyclePhase} columns in \code{colData(sce)}
}
\description{
Calculate the cell cycle phase for every single cell using tricycle
}
\details{
Wrapper to the \code{tricycle} method (Zheng et al. 2022) for cell 
    cycle estimation, which uses a dataset with a continuum, periodic cell 
    cycle and projects the user data onto the embeddings. This wrapper only 
    adds two options for an automatic recentering procedure: \code{"optim"} 
    is an optimization routine to identify the point in the center which 
    is farthest from all other points, and \code{"chull"} calculates the center
    of the convex hull of the 2D embedding.
    
    Moreover, a crude phase assignment is calculated based on the angle (this
    is not part of the original publication but these assignments are detailed
    in the vignette.)
}
