% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIdentities.R
\name{buildReference}
\alias{buildReference}
\title{Build a reference}
\usage{
buildReference(sce, agg_by, agg_assay = "logcounts", agg_fun = "sum")
}
\arguments{
\item{sce}{a SingleCellExperiment object containing the reference data}

\item{agg_by}{character, the name of the \code{colData(sce)} column that the 
aggregation should be done by, e.g. a column with labels or clustering
results}

\item{agg_assay}{character, the name of the \code{assay(sce)} slot whose values
will be aggregated. Default is "logcounts"}

\item{agg_fun}{character, the aggregation function, one of "sum" or "mean". 
Default is "sum".}
}
\value{
a matrix with the same row names as \code{sce} and values aggregated by
    unique values of \code{agg_by}, which will be the column names.
}
\description{
Summarizes an assay from a SingleCellExperiment object based on a character column
}
