% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellula.r
\name{cellula}
\alias{cellula}
\title{#' SCE Integration pipeline}
\usage{
cellula(
  sce,
  batch = NULL,
  name = NULL,
  do_qc = TRUE,
  do_norm = TRUE,
  geneset_list = NULL,
  discard = TRUE,
  subset_mito = TRUE,
  subset_ribo = TRUE,
  subset_malat1 = TRUE,
  detect_doublets = TRUE,
  run_emptydrops = FALSE,
  emptydrops_cutoff = "auto",
  emptydrops_alpha = 0.01,
  hvg_ntop = 2000,
  integration_method = "fastMNN",
  ndims = 20,
  verbose = FALSE,
  save_plots = TRUE,
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{batch}{character, the name of the column in `colData(sce)` with batch labels}

\item{name}{character, the name of the project/folder where files will be saved.
If NULL, a random name will be generated}

\item{do_qc}{logical, should QC steps be performed? Default is TRUE}

\item{do_norm}{logical, should normalization and dimensionality reduction be
performed? Default is TRUE}

\item{geneset_list}{named list of gene IDs (must coincide with rownames of the
SCE object) that will be used by AUCell. Default is NULL meaning no AUC
will be calculated.}

\item{discard}{logical, should values that do not meet QC thresholds be discarded?
Default is TRUE.}

\item{subset_mito}{logical, should mitochondrial transcripts be used for QC?
Default is TRUE.}

\item{subset_ribo}{logical, should ribosomal transcripts be used for QC?
Default is TRUE.}

\item{subset_malat1}{logical, should MALAT1 transcripts be used for QC?
Default is TRUE.}

\item{detect_doublets}{logical, should `scDblFinder` be run? Default is TRUE.}

\item{run_emptydrops}{logical, should `emptyDrops` be run? Default is TRUE.}

\item{emptydrops_cutoff}{either "auto" (default, barcode rank inflection point)
or a numeric. Cells with total reads below this cutoff are used to calculate
ambient RNA profiles and are removed.}

\item{emptydrops_alpha}{numeric, the FDR threshold to call an empty barcode.}

\item{hvg_ntop}{numeric, the number of top highly variable genes to be selected.
Default is 2000.}

\item{integration_method}{character, one of "fastMNN", "Harmony", "Seurat",
"LIGER", or "regression".}

\item{ndims}{numeric, the number of dimensions to retain in the reduced dimension
embedding for downstream applications. Default is 20.}

\item{verbose}{logical, display messages on progress? Default is FALSE.}

\item{save_plots}{logical, should plots be drawn and saved? Default is TRUE}

\item{parallel_param}{a BiocParallel object specifying the parallelization backend
to be used in some steps of the pipeline. Note: for Seurat options, the
`future` framework should be set up with maximum size and number of cores.}
}
\value{
a `SingleCellExperiment` object with normalized data, doublet assignment
   (if calculated), uncorrected and corrected PCA and UMAP coordinates according
   to the method of choice.
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
