% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateSCE.R
\name{diagnoseIntegration}
\alias{diagnoseIntegration}
\title{Integration diagnostics}
\usage{
diagnoseIntegration(sce, batch, dr, ndims = 20, method = "iLISI", verbose, ...)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{batch}{character, the name(s) of the column(s) in \code{colData(sce)} with batch labels.
It can be more than one column.}

\item{dr}{character, the name(s) of the dimensional reduction slot(s) to be tested for integration. 
It can be more than one slot.}

\item{ndims}{numeric, the number of dimensions to use for integration. Default
is 20.}

\item{method}{character, the integration diagnosis method. One of \code{"iLISI"}, \code{"Gini"},
\code{"Jaccard"}, or \code{"HDB"}. Default is \code{"iLISI"}.}

\item{verbose}{logical, display messages on progress? Default is FALSE.}

\item{...}{extra arguments passed to the main diagnosis functions used by each method (see Details).}
}
\value{
depending on the diagnosis methods of choice, a list, a matrix, or a data frame (see Details).
}
\description{
Diagnose the integration or batch effect correction with different metrics
}
\details{

}
