% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEmptyDrops.R
\name{doEmptyDrops}
\alias{doEmptyDrops}
\title{SCE empty droplet removal sub-pipeline}
\usage{
doEmptyDrops(
  sce,
  batch = NULL,
  emptydrops_cutoff = "auto",
  emptydrops_alpha = 0.01,
  verbose = TRUE,
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{batch}{character, the name of the column in `colData(sce)` with batch labels.
Default is NULL meaning no batches will be considered, and data will be
processed as a single batch.}

\item{emptydrops_cutoff}{either "auto" (default, barcode rank inflection point)
or a numeric. Cells with total reads below this cutoff are used to calculate
ambient RNA profiles and are removed.}

\item{emptydrops_alpha}{numeric, the FDR threshold to call an empty barcode.}

\item{verbose}{logical, display messages on progress? Default is FALSE.}

\item{parallel_param}{a BiocParallel object specifying the parallelization backend
to be used in some steps of the pipeline.}
}
\value{
a SingleCellExperiment object without empty droplets as assigned by emptyDrops
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
