% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{doLabelGSEA}
\alias{doLabelGSEA}
\title{Per-label GSEA}
\usage{
doLabelGSEA(dge, pathways, alpha = 0.05)
}
\arguments{
\item{dge}{a list of data frames with the differential expression results
from the \code{doPBDGE} function.}

\item{pathways}{a named list of genesets/pathways to use in the GSEA}

\item{alpha}{numeric, the FDR threshold to use for filtering}
}
\value{
a named list with GSEA results (\code{gsea}), and a matrix (\code{nesmat}) 
	   where each column represents a label DE for which GSEA was performed, and each 
    row a geneset/pathway that is significant in at least one of the labels.
}
\description{
Perform GSEA on a list of per-label DE results
}
