% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doNormAndReduce.R
\name{doNormAndReduce}
\alias{doNormAndReduce}
\title{SCE Normalization and dimensionality reduction sub-pipeline}
\usage{
doNormAndReduce(
  sce,
  batch,
  name,
  ndims = 20,
  hvg_ntop = 2000,
  verbose = TRUE,
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{batch}{character, the name of the column in `colData(sce)` with batch labels}

\item{name}{character, name of the folder and file where results are stored}

\item{ndims}{numeric, number of dimensions (PCs) to be used for UMAP construction.
Default is 20.}

\item{hvg_ntop}{numeric, number of top HVGs to retain}

\item{verbose}{logical, display messages on progress? Default is FALSE.}

\item{parallel_param}{a BiocParallel object specifying the parallelization backend
to be used in some steps of the pipeline.}
}
\value{
a SingleCellExperiment object containing normalized values and a non-integrated
   PCA and UMAP dimensionality reductions.
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
