% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{doPBDGE}
\alias{doPBDGE}
\title{Pseudo-bulk differential expression analysis}
\usage{
doPBDGE(
  sce,
  replicates,
  labels,
  condition,
  contrast = NULL,
  design = NULL,
  verbose = TRUE,
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{replicates}{character, column in \code{colData(sce)}
that contains the replicate information}

\item{labels}{character, column in \code{colData(sce)}
that contains the label information}

\item{condition}{character, column in \code{colData(sce)}
that contains the condition information}

\item{contrast}{character, the contrast to be tested.
Default is the second level of the condition minus the first level.
See Details for how it is formatted.}

\item{design}{a design matrix for the differential expression analysis.
Default is NULL, meaning the design matrix will be created automatically
with an intercept term set to 0, e.g. \code{~0 + condition}.}

\item{verbose}{logical, display messages on progress? Default is \code{TRUE}.}

\item{parallel_param}{a \code{BiocParallel} object specifying
the parallelization backend}
}
\value{
a data frame with the differential expression results
    through the \code{pseudoBulkDGE()} function from \code{{scran}}.
}
\description{
This function performs differential expression
    analysis on a single-cell experiment aggregated into pseudobulk.
}
\details{
This function is a wrapper around the \code{pseudoBulkDGE()}
   function from \code{{scran}}. It requires the user to specify the
   columns in the \code{colData(sce)} that contain the replicate
   information, the label information, and the condition information.

   The aggregation happens by taking the combination of the condition,
   replicate, and label columns as individual pseudobulk samples. The
   differential expression analysis is then performed on these samples
   assuming \code{condition} is the categorical variable to use for DE.
   Any number of covariates can be included in the \code{design} argument.

   The contrast to be tested can be specified as well. If not, the default
   is the second level of the condition minus the first level. The format
   of the contrast is the one accepted by \code{pseudoBulkDGE()}, i.e.
   \code{"conditionLevel1-conditionLevel2"} where condition is the name of
   the column containing the condition, and Level1 and Level2 are the two
   levels to compare.

   Default design is a formula with an intercept term set to 0, e.g.
   \code{~0 + condition} as it forces the user to explicitly declare in the
   contrast which is the reference level. However, if design is provided,
   the user can specify a design without intercept (e.g. \code{~condition})
   which means that the reference level for the factor will be taken as the
   reference level for the contrast too. Not knowing what the reference
   value for the factor is can cause the function to fail if the reference
   level is used in the \code{contrast} argument.
}
