% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cpal_qual_tritan}
\alias{.cpal_qual_tritan}
\title{Qualitative color palette - tritanopy}
\usage{
.cpal_qual_tritan(n)
}
\arguments{
\item{n}{numeric, the number of colors. Must be between 2 and 30 inclusive.}
}
\value{
a color palette with n colors, optimized for tritanopy
}
\description{
Preset qualitative color palettes from qualpalr
}
\details{
These palettes were generated using the \code{qualpalr} package by 
    Johan Larsson, and are optimized for colors that are maximally distant
    in a perceptual color space. They are listed here to avoid a dependency
    but it is important to credit the original author of the code that 
    generated these palettes.
    
    These palettes simulate tritanopy (color vision deficiency) at a severity
    of 0.5.
}
