% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cpal_seq_turbo}
\alias{.cpal_seq_turbo}
\title{Sequential color palette - turbo}
\usage{
.cpal_seq_turbo()
}
\description{
Preset continuous color palettes from pals
}
\details{
These palettes were generated using the \code{pals} package by 
    Kevin Wright.
}
