% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cpal_seq_ylgnbu}
\alias{.cpal_seq_ylgnbu}
\title{Sequential color palette - YlGnBu truncated}
\usage{
.cpal_seq_ylgnbu()
}
\description{
Preset continuous color palettes from colorspace
}
\details{
These palettes were generated using the \code{colorspace} package by 
    Ross Ihaka, Paul Murrell, Kurt Hornik, Hason Fisher, Reto Stauffer, 
    Claus Wilke, Claire McWhite, and Achim Zeileis.
}
