% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsampleCells}
\alias{downsampleCells}
\title{Downsample cells}
\usage{
downsampleCells(
  sce,
  sample_by,
  proportion = 0.1,
  min = 10,
  verbose = TRUE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{sample_by}{character, the column name of \code{colData(sce)} for sampling. 
Cells will be downsampled within each level of \code{sample_by}.}

\item{proportion}{numeric, the target proportion of cells to downsample within
each level of \code{sample_by}.}

\item{min}{numeric, the minimum number of cells to retain within each level of
\code{sample_by}. Strictly positive.}

\item{verbose}{logical, should progress be displayed? Default is \code{TRUE}.}

\item{BPPARAM}{a \code{BiocParallel} \code{BPPARAM} object. 
Default is \code{SerialParam()}, meaning no parallelization will be used.}
}
\value{
a \code{SingleCellExperiment} object with fewer cells than the original
}
\description{
Downsamples a SingleCellExperiment to a proportion of cells
}
\author{
Giuseppe D'Agostino
}
