% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsampleCounts}
\alias{downsampleCounts}
\title{Downsample counts}
\usage{
downsampleCounts(
  mat,
  target = NULL,
  chunksize = 1000,
  BPPARAM = SerialParam(progressbar = verbose),
  verbose = TRUE
)
}
\arguments{
\item{mat}{a matrix where rows are genes/features and columns are samples}

\item{target}{numeric, the total number of counts per sample to be reached by
downsampling}

\item{chunksize}{numeric, the size (in number of samples) to be processed at
a time}

\item{BPPARAM}{a BiocParallel BPPARAM object. Default is SerialParam(), meaning
no parallelization will be used.}

\item{verbose}{logical, should progress be displayed? Default is TRUE.}
}
\value{
a sparse matrix of class "dgCMatrix"
}
\description{
Downsamples a count matrix to a predetermined total count
}
\author{
Giuseppe D'Agostino, inspired by Scott R. Tyler's \code{downsample} package
}
