% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateSCE.R
\name{integrateSCE}
\alias{integrateSCE}
\title{SCE Integration pipeline}
\usage{
integrateSCE(
  sce,
  batch,
  hvgs,
  hvg_ntop = 2000,
  method = "fastMNN",
  ndims = 20,
  neighbor_n = NULL,
  parallel_param = SerialParam(),
  verbose = FALSE
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{batch}{character, the name of the column in \code{colData(sce)} with batch labels}

\item{hvgs}{character, vector of highly variable gene IDs}

\item{hvg_ntop}{numeric, the number of top HVGs to retain. Default is 2000.}

\item{method}{character, the integration method. One of \code{"fastMNN"}, \code{"Harmony"},
\code{"Seurat"}, \code{"LIGER"}, and \code{"regression"}. Default is \code{"fastMNN"}.}

\item{ndims}{numeric, the number of dimensions to use for integration. Default
is 20. For LIGER, it is the value of the k parameter in \code{rliger::optimizeALS()}}

\item{neighbor_n}{the number of neighbors used to compute UMAP. Default is
NULL, which results in the rounded squared root of the number of cells.}

\item{parallel_param}{a \code{BiocParallel} object specifying the parallelization backend
to be used in some steps of the pipeline. Note: for Seurat options, the
\code{future} framework should be set up with maximum size and number of cores.}

\item{verbose}{logical, display messages on progress? Default is FALSE.}
}
\value{
a `SingleCellExperiment` object with integrated dimensionality reduction.
    In the case of LIGER integration, only the `H.norm` slot will be returned.
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
