% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphsAndClusters.R
\name{makeGraphsAndClusters}
\alias{makeGraphsAndClusters}
\title{SCE multi-resolution clustering}
\usage{
makeGraphsAndClusters(
  sce,
  neighbors = 10L,
  weighting_scheme = "jaccard",
  sweep_on = "clustering",
  method = "louvain",
  k = seq(0.1, 1, length.out = 6),
  space = NULL,
  ndims = 20L,
  calculate_modularity = TRUE,
  calculate_silhouette = TRUE,
  leiden_iterations = 5L,
  save_graphs = FALSE,
  prefix = "SNN_",
  verbose = FALSE
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{neighbors}{numeric, number of neighbors for SNN graph edge construction.
Default is 10.}

\item{weighting_scheme}{character, the weighting scheme for SNN graph construction.
One of \code{"jaccard"}, \code{"rank"}, \code{"number"}. Default is \code{"jaccard"}.}

\item{sweep_on}{character, the parameter used for sweeping. Can be \code{"clustering"},
meaning values of \code{k} will be looped through as resolution, or \code{"SNN"},
meaning values of \code{k} will be looped through as number of neighbors.}

\item{method}{character, the type of graph-based clustering to use. One of
\code{"louvain"} or \code{"leiden"}. Default is \code{"louvain"}.}

\item{k}{numeric, vector of parameter sweep for graph construction or clustering.}

\item{space}{a matrix of lower dimensional embedding such as the PCA coordinates.
if \code{NULL} (default), the \code{"PCA"} slot from \code{reducedDims(sce)}.}

\item{ndims}{numeric, the number of dimensions (columns of \code{space}) to use to
build the SNN graph. Default is 20}

\item{calculate_modularity}{logical, should pairwise modularity between
clusters be calculated? Default is \code{TRUE}}

\item{calculate_silhouette}{logical, should approximate silhouette widths be
calculated? Default is \code{TRUE}}

\item{leiden_iterations}{numeric, the number of iterations of Leiden clustering.
Default is 5.}

\item{save_graphs}{logical, should the \code{igraph} graph objects be saved in
the metadata of \code{sce}? Default is FALSE.}

\item{prefix}{character, the prefix of the column names on \code{colData(sce)}
where clustering results are stored. Default is \code{SNN_}.}

\item{verbose}{logical, should messages be written? Default is \code{FALSE}}
}
\value{
a \code{SingleCellExperiment} object with cluster memberships in the
    \code{colData} table, named according to prefix and respective value of \code{k}.
    Optionally, silhouette and/or modularity values are stored in the \code{metadata}
    slot of the \code{SingleCellExperiment} object, one for every value of \code{k}.
}
\description{
Very simple wrapper to SNN graph and Louvain/Leiden clustering using
multiple resolutions
}
