% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{makeNESmatrix}
\alias{makeNESmatrix}
\title{GSEA NES matrix}
\usage{
makeNESmatrix(gsea_list, alpha = 0.05)
}
\arguments{
\item{gsea_list}{a list of GSEA results from \code{link{doLabelGSEA}}.}

\item{alpha}{numeric, the FDR threshold to use for filtering}
}
\value{
a matrix where each column represents a label DE 
    for which GSEA was performed, and each row a geneset/pathway 
    that is significant in at least one of the labels.
}
\description{
Create a NES matrix from a list of GSEA results
}
