% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dots}
\alias{plot_dots}
\title{Plot a dot plot}
\usage{
plot_dots(
  sce,
  genes,
  group_by,
  cluster_genes = TRUE,
  cluster_groups = TRUE,
  exprs_use = "logcounts",
  threshold = 0,
  color_palette = NULL,
  format = "wide"
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{genes}{character string, the genes to be plotted (matched with rownames)}

\item{group_by}{character, column name in the colData slot of the SCE object, 
e.g. "cluster". Will be used to assign calculate proportions. 
Must be categorical (factor or coercible character).}

\item{cluster_genes}{logical, should genes be clustered? Default is TRUE}

\item{cluster_groups}{logical, should groups be clustered? Default is TRUE}

\item{exprs_use}{character, name of the `assay` slot in the SingleCellExperiment}

\item{color_palette}{a character string containing colors to be used. Default
is NULL, meaning an automatic palette will be generated.}

\item{format}{character, one of "wide" (genes are columns, groups are rows) 
or "tall" (genes are rows, grups are columns)}
}
\value{
a ggplot object showing average gene expression and proportion of 
   expressing cells grouped by a grouping variable such as cluster.
}
\description{
Plots a dot plot with gene expression from a SingleCellExperiment object
}
