% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{Plot UMAP}
\usage{
plot_UMAP(...)
}
\arguments{
\item{...}{see \code{\link{plot_DR}}}
}
\value{
see \code{\link{plot_DR}}
}
\description{
Alias for plot_DR for consistency with previous iterations
}
