% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotClusTree}
\alias{plotClusTree}
\title{Plot clustering tree}
\usage{
plotClusTree(
  sce,
  prefix = "SNN_",
  additional_labels = NULL,
  color_palette = NULL
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{prefix}{character, the prefix of the clustering labels. Default is \code{"SNN_"}}

\item{additional_labels}{character, additional labels to include in the tree. Default is \code{NULL}}

\item{color_palette}{named character vector, the color palette to use for the clusters. Default is \code{NULL}}
}
\value{
a \code{ggraph} object with the clustering tree
}
\description{
Plot a clustering tree using clustering or any other labels
}
\details{
the function generates a tree of the clustering labels as implemented 
    in the \code{clustree} package by Zappia and Oshlack (2018) (\url{https://doi.org/10.1093/gigascience/giy083}). 
	   
The function generates a graph of the clustering labels using the \code{igraph}
    package, and plots it using the \code{ggraph} package. 

	   This is a much simpler implementation of the principle behind \code{clustree}, with fewer options and 
    possibilities compared to the original package.

    The nodes are colored by the resolution and/or label parameter, and the size of 
    the nodes is proportional to the number of cells in each cluster. The edges are colored 
    by the number of cells in each cluster that overlap between two consecutive clustering 
    resolutions, and shaded by the proportion of overlapping cells. 

    Groupings identified by the prefix are assumed to have a numeric resolution value 
    after the prefix, and the function will sort these grouping by this value.

    Additional labels (specified in the \code{additional_labels} argument) will be included in the tree 
    in the order in which they were provided.
}
