% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle.R
\name{plotCycle}
\alias{plotCycle}
\title{Plot cell cycle}
\usage{
plotCycle(sce, rings_by = NULL, color_by = "tricyclePosition")
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{rings_by}{character, column in \code{colData(sce)} to divide the embedding by}

\item{color_by}{character, column in \code{colData(sce)}. Default is \code{"tricyclePosition"}.}
}
\value{
a plot of cell cycle assignments
}
\description{
Plots the cell cycle assignment in a circular embedding
}
