% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDEHeatmap}
\alias{plotDEHeatmap}
\title{Differential gene expression heatmap}
\usage{
plotDEHeatmap(
  sce,
  genes,
  dge,
  exprs = "logcounts",
  alpha_include = 1,
  alpha_dot = 0.05,
  condition_pal = NULL,
  cluster_pal = NULL,
  title = NA,
  cluster_cols = FALSE,
  gaps = TRUE,
  scale = TRUE
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{genes}{character, the genes to include in the heatmap.
These will be looked up in the rownames of the object, which are also the 
   rownames of the data frames in the DE results. At least 2 genes must be provided.}

\item{dge}{a data frame with the differential expression results from \code{\link{doPBDGE}}}

\item{exprs}{character, the expression values to use. Default is \code{"logcounts"}}

\item{alpha_include}{numeric, the FDR threshold to include a gene in the heatmap
Default is 1 (all genes included)}

\item{alpha_dot}{numeric, the FDR threshold to mark a gene as significant in
the heatmap using a dot. Default is 0.05}

\item{condition_pal}{a named vector with the colors for the conditions. Default is NULL}

\item{cluster_pal}{a named vector with the colors for the clusters. Default is NULL}

\item{title}{character, the title of the heatmap. Default is NA, no title will be displayed.}

\item{cluster_cols}{logical, should the columns be clustered? Default is \code{TRUE}}

\item{gaps}{logical, should gaps be included between the columns? Default is \code{FALSE}}

\item{scale}{logical, should the expression data be scaled by row? Default is \code{TRUE}}
}
\value{
a composite heatmap of the pseudo-bulk aggregated genes with their DE results
}
\description{
Draws a heatmap of aggregated genes together with their DE results
}
\details{
this function generates a heatmap of genes using pseudo-bulk aggregated profiles
   taken by aggregating the \code{sce} object by the variables stored in the attributes
   of the \code{dge} object, and taking the assay specified by \code{exprs}. 
   Then, a second heatmap is drawn on the right hand side of the first heatmap, showing
   the DE results for the contrast stored in the attributes of the \code{dge} object.
		
   The genes are filtered by the \code{alpha_include} threshold, and significant genes
   are labelled in the right hand side heatmap using a white dot if their FDR is below
	  \code{alpha_dot}.
}
