% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDGEStripchart}
\alias{plotDGEStripchart}
\title{Differential gene expression stripchart}
\usage{
plotDGEStripchart(dge, alpha = 0.05, lfc = 1, title = NULL)
}
\arguments{
\item{dge}{a data frame with the differential expression results
from the \code{doPBDGE} function}

\item{alpha}{numeric, the FDR threshold to use for filtering}

\item{lfc}{numeric, the log-fold change threshold to use for filtering}

\item{title}{character, the title of the plot}
}
\value{
a stripchart with per-cluster DE results
}
\description{
Plots a stripchart of the differential gene expression results
}
