% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Gene expression heatmap}
\usage{
plotHeatmap(
  sce,
  genes,
  coldata_cols = NULL,
  scale = TRUE,
  exprs = "logcounts",
  aggregate = FALSE,
  aggregate_by = NULL,
  color_pal = NULL,
  annotation_pal = NULL,
  cluster_cells = TRUE,
  cluster_genes = TRUE,
  gaps = NULL,
  clip_values = 99.5,
  raster = FALSE,
  ...
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{genes}{character, the genes to include in the heatmap.
These will be looked up in the rownames of the object, which are also the 
   rownames of the data frames in the DE results. At least 2 genes must be provided.}

\item{coldata_cols}{character, the columns in the \code{colData} slot of the \code{sce} object
to include in the heatmap as annotations. Their order influences column ordering
   when \code{cluster_cols = FALSE}. Default is \code{NULL}.}

\item{scale}{logical, should the expression data be scaled by row? Default is \code{TRUE}}

\item{exprs}{character, the expression values to use. Default is \code{"logcounts"}}

\item{aggregate}{logical, should the data be aggregated? If \code{TRUE}, the data will be
aggregated by the variables indicdated in \code{aggregate_by}. Default is \code{FALSE}}

\item{aggregate_by}{character, the columns in the \code{colData} slot of the \code{sce} object
to aggregate by. Default is \code{NULL}}

\item{color_pal}{character, the color palette to use for the heatmap. Default is \code{NULL}}

\item{annotation_pal}{named list of named character vectors, the color palette 
to use for the annotations. Default is \code{NULL}. See Details}

\item{cluster_cells}{logical, should the columns (cells) be clustered? 
Default is \code{TRUE}}

\item{cluster_genes}{logical, should the rows (genes) be clustered? Default is \code{TRUE}}

\item{gaps}{character, the columns in \code{coldata_cols} to leave gaps between. See Details
Default is \code{NULL}.}

\item{clip_values}{numeric, the percentile to clip the values at. Default is \code{99.5}}

\item{raster}{logical, should the heatmap be rasterized? Default is \code{FALSE}, but 
it is changed to \code{TRUE} automatically if the number of cells or genes is 
   greater than 2000. This behaviour can be overriden by setting \code{raster = "force"}.}

\item{...}{additional arguments to pass to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
a heatmap of the gene expression data, optionally with annotations 
   from the \code{colData} slot
}
\description{
Draws a heatmap of gene expression from a SingleCellExperiment object
}
\details{
this function generates a heatmap of genes using expression data from the \code{sce} 
	  object (specified by \code{exprs}) with optional aggregation. The heatmap can be annotated
	  with any number of column annotations, i.e. colouring segments above each column according
   to the \code{colData} slot of the \code{sce} object. 
		
	  The \code{annotation_pal} argument is a named list of named character (colour) vectors, 
	  where each element of the list corresponds to a column in \code{coldata_cols}. 
	  The named character vectors should have the same length as the number of unique values 
   in the column, and the names should be the unique values. 
	
	  The order of colData columns in \code{coldata_cols} influences the order of the columns in the
	  heatmap. If \code{cluster_cols = FALSE}, the columns will be ordered according to the order
	  of the columns in \code{coldata_cols}, where the first element dictates the major order, then
	  the second element the minor order, etc, e.g. if \code{coldata_cols = c("condition", "label")}
   columns will be first ordered by condition and then, within each condition, by label.
	  If \code{cluster_cols = TRUE}, the columns will be ordered according to the clustering of the columns.
}
