% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLabelMD}
\alias{plotLabelMD}
\title{Per-label Mean-Difference plot}
\usage{
plotLabelMD(dge, ntop = 5, alpha = 0.05, lfc = 1)
}
\arguments{
\item{dge}{a data frame with the differential expression results
from the \code{doPBDGE} function}

\item{ntop}{numeric, the number of top genes to show in the plot}

\item{alpha}{numeric, the FDR threshold to use for filtering}

\item{lfc}{numeric, the log-fold change threshold to use for filtering}
}
\value{
a faceted plot where each panel shows the mean (log(CPM)) vs logFC 
   for all genes in a label that do not have NA DE test results.
}
\description{
Plot a Mean-Difference plot of the DE results per label
}
