% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{plotLabelPCA}
\alias{plotLabelPCA}
\title{Per-label PCA plot}
\usage{
plotLabelPCA(sce, replicates, labels, condition)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{replicates}{character, column in \code{colData(sce)}
that contains the replicate information}

\item{labels}{character, column in \code{colData(sce)}
that contains the label information}

\item{condition}{character, column in \code{colData(sce)}
that contains the condition information}
}
\value{
a PC1 vs PC2 faceted plot for pseudobulk profiles 
    where each panel corresponds to a label
}
\description{
Plot a PCA plot of the aggregated data per label using
the first 2 components and top 1000 HVGs
}
