% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metromap.R
\name{plotMetroMap}
\alias{plotMetroMap}
\title{Plot Metro Map}
\usage{
plotMetroMap(
  sce,
  coords_by,
  dimred,
  snap = TRUE,
  res = 30,
  overlay = TRUE,
  overlay_res = 60,
  include_cells = FALSE,
  labels = FALSE
)
}
\arguments{
\item{sce}{a `SingleCellExperiment` class object}

\item{coords_by}{character, a grouping variable to define the clusters.}

\item{dimred}{character, the name of the 2D dimensional reduction to use as
base for plotting}

\item{snap}{logical, should points and lines be snapped to a regular grid?
default is TRUE.}

\item{res}{numeric, the resolution of the grid for snapping. Higher = smaller
spacing between grid points, and more points. Default is 30.}

\item{overlay}{logical, should an overlay contour be drawn? Default is TRUE}

\item{overlay_res}{the overlay resolution. Default is 60.}

\item{include_cells}{logical, should cells be plotted in the background? 
Default is FALSE.}

\item{labels}{logical, should cluster labels be plotted? Default is TRUE.}
}
\value{
A plot where lineages are represented as metro lines.
}
\description{
Plots a metro map-like representation of trajectories
}
