% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_clustering_slingshot.r
\name{plotModularity}
\alias{plotModularity}
\title{Plot cluster modularity}
\usage{
plotModularity(sce, name)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{name}{character, the name in the metadata slot of the SCE object, e.g.
"modularity_SNN_100"}
}
\value{
a heatmap of pairwise modularity
}
\description{
Plots a heatmap showing pairwise cluster modularity from a SCE object
}
