% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProportions}
\alias{plotProportions}
\title{Plot proportions}
\usage{
plotProportions(sce, label, group_by, color_palette = NULL)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{label}{character, the column name in the
colData of the SingleCellExperiment used for
colouring the plot}

\item{group_by}{character, the column name in the
colData of the SingleCellExperiment used for
grouping the bars}

\item{color_palette}{named character, the name of the color
palette to be used. Default is \code{NULL},
meaning an automatic palette will be generated.}
}
\value{
a stacked bar plot of proportions of cells in each
   label, grouped according to the \code{group_by} variable.
}
\description{
Plots the proportion of cells in each label and grouping
}
