% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPseudotimeHeatmap}
\alias{plotPseudotimeHeatmap}
\title{Pseudotime heatmap}
\usage{
plotPseudotimeHeatmap(
  sce,
  genes,
  pseudotime,
  exprs = "logcounts",
  cluster_genes = FALSE,
  labels = NULL,
  labels_pal = NULL,
  smoothing_window = 200,
  raster = FALSE
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{genes}{character, the genes to include in the heatmap.}

\item{pseudotime}{character, the column in the \code{colData} 
slot of the \code{sce} object with pseudotime values.}

\item{exprs}{character, the name of the assay with expression values to use. 
Default is \code{"logcounts"}}

\item{cluster_genes}{logical, should the genes be clustered? 
Default is \code{FALSE}}

\item{labels}{character, the column in the \code{colData} slot of the \code{sce} object
to use as labels for heatmap annotation. Default is \code{NULL}}

\item{labels_pal}{named list of named character vectors, the color palette
to use for the labels. Default is \code{NULL}}

\item{smoothing_window}{numeric, the window size for smoothing the expression values.
Default is \code{200}}

\item{raster}{logical, should the heatmap be rasterized? Default is \code{FALSE}}
}
\value{
a heatmap of the gene expression data ordered by pseudotime with,
	   optionally, the labels of each cells as annotation on top of the heatmap.
	   Each gene is scaled between 0 and 1 and smoothed using a zero-padded moving 
	   average with the specified window size.
}
\description{
Draws a heatmap of pseudotime-ordered gene expression
}
