% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotROC}
\alias{plotROC}
\title{plot Receiver-Operator Characteristic curves}
\usage{
plotROC(sce, gene, label, exprs = "logcounts", plot = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{gene}{character, the gene to use for the ROC curve}

\item{label}{character, the column in the \code{colData} slot of the \code{sce} object
to use as labels for classification and facetting}

\item{exprs}{character, the name of the assay with expression values to use.
Default is \code{"logcounts"}}

\item{plot}{logical, should the ROC curves be plotted? Default is \code{TRUE}}
}
\value{
if \code{plot = TRUE}, a facetted plot with ROC curves 
   for each label in the \code{label} column together with other 
   classification statistics (see details). 
	  If \code{plot = FALSE}, a list with the ROC curve data and other
   statistics is returned.
}
\description{
plot ROC curves with other classification statistics
}
\details{
the function calculates the ROC curve for the gene specified 
    in the \code{gene} argument, showing the classification performance 
	   of the gene for the label against all other labels. The function
	   calculates the Area Under the Curve (AUC), Youden's J statistic,
	   the entropy of the gene expression distribution, and the Jensen-Shannon
	   Divergence between the gene expression distribution and an idealized 
    distribution where the gene is only expressed in one label.
}
