% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot approximate silhouette widths}
\usage{
plotSilhouette(sce, name)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{name}{character, the name in the metadata slot of the \code{SingleCellExperiment} object, e.g.
"silhouette_SNN_100"}
}
\value{
a beeswarm plot of silhouette widths
}
\description{
Plots a beesswarm plot showing the approximate width of each cell within
each cluster
}
\details{
This plotting function has been directly taken from the OSCA book
    ("Orchestrating Single Cell Analysis with Bioconductor").
}
