% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stacked_Violin}
\alias{stacked_Violin}
\title{Plot a stacked violin plot}
\usage{
stacked_Violin(
  sce,
  features,
  cluster,
  split_by = NULL,
  stack = TRUE,
  de = NULL,
  significance = FALSE,
  color_palette = NULL,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{features}{character vector, list of features from \code{sce}}

\item{cluster}{character, column name in the \code{colData} slot of \code{sce},
e.g. "cluster". Will be used to assign colors. 
Must be a categorical variable.}

\item{split_by}{character, column name in the \code{colData} slot of the \code{sce},
e.g. "individual". Will be used to divide violin plots. 
Must be a categorical variable.}

\item{stack}{logical, should all violin plots be stacked on 1 column? 
Default is TRUE.}

\item{de}{(optional) data frame containing the results of differential expression.
Standardized column names are enforced.}

\item{significance}{logical, should a bar with adjusted p-value and log2(FC) values
be plotted on top of the violin plots? Default is FALSE. Requires \code{de} to
be specified as well.}

\item{color_palette}{character, a vector of colors to be used for the \code{fill}
aesthetic for the violin plot areas.}

\item{nrow}{numeric, the number of rows to build the facetted plot. Only used
if \code{stacked = FALSE}.}

\item{ncol}{numeric, the number of columns to build the facetted plot. Only used
if \code{stacked = FALSE}.}
}
\value{
a ggplot object showing a facetted violin plot with several features grouped
    and colored according to two categorical variables.
}
\description{
Plots a stacked violin plot of at least two features in different groupings
}
\details{
This function has two possible outcomes, controlled by the \code{stack} argument: 
    1) \code{stack = TRUE} returns a stacked violin plot where each feature is a row, and 
    violins correspond to whatever grouping is specified in "cluster"; 
    2) \code{stack = FALSE} returns a faceted grid of violin plots where each feature is a
    panel.

    The first behaviour is useful to check markers/gene expression patterns in different 
    groupings with an idea of the distribution. It reproduces \code{scanpy}'s stacked
    violin plot. 
    The second behaviour is useful to show comparisons between groupings/conditions 
    for a set of features, and allows differential expression significance information
    to be overlaid. 

    Since the idea is to make different groupings and features comparable
    the \code{\link[ggplot2](geom_violin)} geom is set with \code{scale = "width"}, which can show the
    distribution of the data but does not scale with the number of data points (cells).
}
