% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignIdentities.R
\name{assignIdentities}
\alias{assignIdentities}
\title{SCE Cell Type annotation pipeline}
\usage{
assignIdentities(
  sce,
  genesets = NULL,
  method,
  ref = NULL,
  assay = NULL,
  verbose = TRUE,
  name = NULL,
  return_scores = FALSE,
  kcdf = "Gaussian",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{genesets}{named list of character vectors, genesets used to calculate
labels and/or assign scores. Alternatively, a character vector to calculate
a single score from.}

\item{method}{character, the method for scoring. One of \code{"AUC"}, \code{"Seurat"},
\code{"ssGSEA"}, \code{"UCell"}, or \code{"Jaitin"}.}

\item{ref}{a matrix containing expression values from reference transcriptomes. 
Only used when \code{method = "Jaitin"}.}

\item{assay}{the name of the slot in \code{sce} containing expression values 
to calculate the log-likelihood. Only used when \code{method = "Jaitin"} or
\code{method = "ssGSEA"}.}

\item{verbose}{logical, should messages on progress be printed? Default is TRUE}

\item{name}{character, the name of the column in \code{colData(sce)} where final
labels will be stored}

\item{return_scores}{logical, should the scores for each cell and each geneset
be saved in the object metadata slot? Default is \code{FALSE}.}

\item{kcdf}{character, which kernel to use for the CDF. One of \code{"Poisson"} or
\code{"Gaussian"}. Only used when \code{method = "ssGSEA"}.}

\item{BPPARAM}{a \code{BiocParallel} BPPARAM specifying the parallelization. 
Only used when \code{method = "AUC"}. Default is \code{SerialParam()}}

\item{...}{other arguments passed internally to \code{\link[AUCell]{AUCell_calcAUC}}
("AUC" method), \code{\link[Seurat]{AddModuleScore}} ("Seurat" method),
\code{\link[GSVA]{gsva}}("ssGSEA" method), or \code{\link[UCell]{ScoreSignatures_UCell}}
("UCell" method).}
}
\value{
a \code{SingleCellExperiment} object with a column named \code{"name"} containing
    the highest scoring label for a method. Optionally, the single scores from
    each method for each geneset are saved in the \code{metadata} slot.
}
\description{
Automatic assignment of cell type identities in SingleCellExperiment objects
}
\details{
This is a wrapper around four methods for assigning to each cell a
   score for the expression of a given geneset, and a method that uses a reference
   transcriptome as prior probability distribution to estimate the posterior 
   probability that a cell is similar to the reference via a multinomial model.
   \itemize{
    \item{ The "AUC" method
    (\href{https://www.nature.com/articles/nmeth.4463}{Aibar et al. 2017}) calculates
    the Area Under the Curve of the ranked expression of genes in each geneset.}

   \item{The "Seurat" method (\href{https://www.science.org/doi/10.1126/science.aad0501}{Tirosh et al. 2016})
    uses the \code{Seurat::AddModuleScore()} function. }

   \item{The "ssGSEA" method (\href{https://www.nature.com/articles/nature08460}{Barbie et al. 2009})
    calculates an enrichment score in each cell using "single sample" GSEA.}

   \item{The "UCell" method (\href{https://www.sciencedirect.com/science/article/pii/S2001037021002816}{Andreatta and Carmona 2020})
    calculates an enrichment score using a Mann-Whitney U-test.}
    
   \item{The "Jaitin" method (\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4412462/}{Jaitin et al. 2016}) calculates the posterior probability of a single cell being close to a 
    reference transcriptome.}
   }

   If a list of genesets is given, all methods result in a final label assignment
   (i.e. the names of the user-provided geneset for which each single cell has
   the highest score). If only a single geneset is provided, the score for that
   geneset is added to the \code{colData} slot.
   
   For the \code{"Jaitin"} method, the \code{ref} object is a matrix containing
   genes in each row, and samples in each column. It is important to specify the
   \code{assay} slot depending on what the reference contains. If the reference
   contains raw counts, it is advisable to use \code{assay = "counts"}; if it 
   contains log-normalized counts, it is advisable to use \code{assay = "logcounts"}.
   The resulting scores, optionally saved in the \code{metadata}, are the posterior
   probabilities. 
   
   It is possible to use \code{\link{buildReference}} to create a reference matrix
   out of a \code{SingleCellExperiment} object to be used as an input.

   Please note that scores are not comparable between themselves - AUC,
   ssGSEA, UCell and Jaitin are strictly positive, whereas the Module Score can also
   be negative; UCell, AUCell and Jaitin are normalized between 0 and 1, whereas
   ssGSEA is not.

   The user can pass additional arguments to each function using the `...`
   argument.
}
