% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellula.r
\name{cellula}
\alias{cellula}
\title{SingleCellExperiment processing pipeline}
\usage{
cellula(
  sce,
  batch = NULL,
  name = NULL,
  do_qc = TRUE,
  do_norm = TRUE,
  geneset_list = NULL,
  discard = TRUE,
  subset_mito = TRUE,
  subset_ribo = TRUE,
  subset_malat1 = TRUE,
  detect_doublets = TRUE,
  run_emptydrops = FALSE,
  emptydrops_cutoff = "auto",
  emptydrops_alpha = 0.01,
  hvg_ntop = 2000,
  integration_method = "fastMNN",
  ndims = 20,
  verbose = FALSE,
  save_plots = TRUE,
  save_temp = FALSE,
  path = "./",
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{batch}{character, the name of the column in \code{colData(sce)} with batch labels}

\item{name}{character, the name of the project/folder where files will be saved.
If \code{NULL}, a random name will be generated}

\item{do_qc}{logical, should QC steps be performed? Default is \code{TRUE}}

\item{do_norm}{logical, should normalization and dimensionality reduction be
performed? Default is \code{TRUE}}

\item{geneset_list}{named list of gene IDs (must coincide with rownames of
\code{sce}) that will be used by \code{AUCell}. Default is \code{NULL} meaning no AUC
will be calculated.}

\item{discard}{logical, should values that do not meet QC thresholds be discarded?
Default is \code{TRUE}.}

\item{subset_mito}{logical, should mitochondrial transcripts be used for QC?
Default is \code{TRUE}}

\item{subset_ribo}{logical, should ribosomal transcripts be used for QC?
Default is \code{TRUE}}

\item{subset_malat1}{logical, should MALAT1 transcripts be used for QC?
Default is \code{TRUE}}

\item{detect_doublets}{logical, should \code{scDblFinder} be run? Default is \code{TRUE}}

\item{run_emptydrops}{logical, should \code{emptyDrops} be run? Default is \code{FALSE}.}

\item{emptydrops_cutoff}{either \code{"auto"} (default, barcode rank inflection point)
or a numeric. Cells with total reads below this cutoff are used to calculate
ambient RNA profiles and are removed.}

\item{emptydrops_alpha}{numeric, the FDR threshold to call an empty barcode.}

\item{hvg_ntop}{numeric, the number of top highly variable genes to be selected.
Default is 2000.}

\item{integration_method}{character, one of \code{"fastMNN"}, \code{"Harmony"}, \code{"Seurat"},
\code{"LIGER"}, or \code{"regression"}. See \code{?integrateSCE} for more 
information.}

\item{ndims}{numeric, the number of dimensions to retain in the reduced dimension
embedding for downstream applications. Default is 20.}

\item{verbose}{logical, display messages on progress? Default is \code{FALSE}.}

\item{save_plots}{logical, should plots be drawn and saved? Default is \code{TRUE}}

\item{save_temp}{logical, should temporary files be saved as the pipeline progresses?
Default is \code{FALSE}}

\item{path}{character, the path where the files will be saved. Default is the current
working directory.}

\item{parallel_param}{a \code{BiocParallel} object specifying the parallelization backend
to be used in some steps of the pipeline. Default is \code{SerialParam()}
meaning no parallelization will be used. Note: for Seurat options, the
\code{future} framework should be set up with maximum size and number of cores.}
}
\value{
a \code{SingleCellExperiment} object with normalized data, doublet assignment
   (if calculated), uncorrected and corrected PCA and UMAP coordinates according
   to the method of choice.
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
