% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doNormAndReduce.R
\name{doNormAndReduce}
\alias{doNormAndReduce}
\title{SCE Normalization and dimensionality reduction sub-pipeline}
\usage{
doNormAndReduce(
  sce,
  batch = NULL,
  name,
  ndims = 20,
  hvg_ntop = 2000,
  verbose = TRUE,
  parallel_param = SerialParam()
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{batch}{character, the name of the column in \code{colData(sce)} with batch labels.
Default is \code{NULL} meaning no batches will be considered, and data will be
processed as a single batch.}

\item{name}{character, name of the folder and file where results are stored}

\item{ndims}{numeric, number of dimensions (PCs) to be used for UMAP construction.
Default is 20.}

\item{hvg_ntop}{numeric, number of top HVGs to retain}

\item{verbose}{logical, display messages on progress? Default is \code{FALSE}.}

\item{parallel_param}{a \code{BiocParallel} object specifying the parallelization backend
to be used in some steps of the pipeline. Default is \code{SerialParam()},
meaning no parallelization will be used.}
}
\value{
a \code{SingleCellExperiment} object containing normalized values and a non-integrated
   \code{PCA} and \code{UMAP} dimensionality reductions.
}
\description{
Pipeline for automatic processing and integration of SingleCellExperiment objects
}
