% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsampleCounts}
\alias{downsampleCounts}
\title{Downsample counts}
\usage{
downsampleCounts(
  sce,
  target = NULL,
  chunksize = 1000,
  BPPARAM = SerialParam(progressbar = verbose),
  verbose = TRUE
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object from which the \code{counts}
assay will be taken.}

\item{target}{numeric, the total number of counts per sample to be reached by
downsampling}

\item{chunksize}{numeric, the size (in number of samples) to be processed at
a time}

\item{BPPARAM}{a \code{BiocParallel} \code{BPPARAM} object. 
Default is \code{SerialParam()}, meaning no parallelization will be used.}

\item{verbose}{logical, should progress be displayed? Default is \code{TRUE}.}
}
\value{
a sparse matrix of class \code{"dgCMatrix"} with sampled counts
}
\description{
Downsamples a count matrix to a predetermined total count
}
\author{
Giuseppe D'Agostino, inspired by Scott R. Tyler's \code{downsample} package
}
