% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{findTrajectories}
\alias{findTrajectories}
\title{Find trajectories}
\usage{
findTrajectories(
  sce,
  space = "PCA",
  clusters,
  method = "slingshot",
  ndims = 20,
  dr_embed = NULL,
  start = "auto",
  omega = TRUE,
  omega_scale = 1.5,
  do_de = FALSE,
  batch_de = NULL,
  verbose = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{sce}{a `SingleCellExperiment` object}

\item{space}{character, the name of the `reducedDim` slot to use for trajectory
estimation. Default is "PCA".}

\item{clusters}{character, the name of the `colData` column with cluster label
indication.}

\item{method}{character, the method for estimation. Currently only "slingshot".}

\item{ndims}{numeric, the number of dimensions to use in `space`. If the number
of columns in `space` is < `ndims`, it will be used.}

\item{dr_embed}{character, the name of the `reducedDim` slot where curves should
be embedded for plotting. Default is NULL, meaning no embedding will be
performed.}

\item{start}{character, the name of the cluster to be used as starting point.
Default is "auto", implying an entropy-based method will be used to guess
the best starting point.}

\item{do_de}{logical. Should differential expression across trajectories be
performed? Default is FALSE.}

\item{batch_de}{character, the name of the `colData` column to be used as a
blocking factor in the differential expression analysis. Default is NULL.}

\item{BPPARAM}{a `BiocParallelParam` object. Default is NULL.}
}
\value{
a SingleCellExperiment object with slingshot outputs
}
\description{
Finds pseudo-temporal trajectories in a reduced dimensional space
}
