% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGraphsAndClusters.R
\name{makeGraphsAndClusters}
\alias{makeGraphsAndClusters}
\title{SCE multi-resolution clustering}
\usage{
makeGraphsAndClusters(
  sce,
  neighbors = 10L,
  weighting_scheme = "jaccard",
  sweep_on = "clustering",
  method = "louvain",
  k = seq(0.1, 1, length.out = 6),
  space = NULL,
  ndims = 20L,
  calculate_modularity = TRUE,
  calculate_silhouette = TRUE,
  leiden_iterations = 5L,
  prefix = "SNN_",
  verbose = FALSE
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{neighbors}{numeric, number of neighbors for SNN graph edge construction.
Default is 10.}

\item{weighting_scheme}{character, the weighting scheme for SNN graph construction.
One of "jaccard", "rank", "number". Default is "jaccard".}

\item{sweep_on}{character, the parameter used for sweeping. Can be "clustering",
meaning values of `k` will be looped through as resolution, or "SNN",
meaning values of `k`will be looped through as number of neighbors.}

\item{method}{character, the type of graph-based clustering to use. One of
"louvain" or leiden". Default is "louvain".}

\item{k}{numeric, vector of parameter sweep for graph cosntruction or clustering.}

\item{space}{a matrix of lower dimensional embedding such as the PCA coordinates.
if NULL (default), the "PCA" slot from `reducedDims(sce)`}

\item{ndims}{numeric, the number of dimensions (columns of `space`) to use to
build the SNN graph. Default is 20}

\item{calculate_modularity}{logical, should pairwise modularity between
clusters be calculated? Default is TRUE}

\item{calculate_silhouette}{logical, should approximate silhouette widths be
calculated? Default is TRUE}

\item{leiden_iterations}{numeric, the number of iterations of Leiden clustering.
Default is 5.}

\item{prefix}{character, the prefix of the column names on `colData`
where clustering results are stored. Default is "SNN_".}

\item{verbose}{logical, should messages be written? Default is FALSE}
}
\value{
a `SingleCellExperiment` object with cluster memberships in the
    `colData` table, named according to prefix and respective value of `k`.
    Optionally, silhouette and/or modularity values are stored in the `metadata`
    slot of the `SingleCellExperiment` object, one for every value of `k`.
}
\description{
Very simple wrapper to SNN graph and Louvain/Leiden clustering using
multiple resolutions
}
