% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_Coldata}
\alias{plot_Coldata}
\title{Plot cell metadata}
\usage{
plot_Coldata(
  sce,
  y,
  x = NULL,
  color_by = NULL,
  group_by = NULL,
  color_palette = NULL,
  contour = TRUE,
  clustered = TRUE,
  plot_cells = FALSE
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{y}{character, the column from colData(sce) whose values need to be plotted}

\item{x}{character, the column from colData(sce) to be used as a grouping variable
default is NULL, which means all points belong to one violin plot only}

\item{color_by}{character, the column name from colData(sce) to be used as a
colouring variable. If NULL (default), violins (and points) will not be colored.}

\item{group_by}{character, the column name from colData(sce) to be used as a
facetting variable. Default is NULL.}

\item{color_palette}{character string of colors for categorical plots or heatmap.
Default is NULL.}

\item{contour}{logical, should contours be plotted on the scatterplot?
default is TRUE.}

\item{clustered}{logical, should rows and columns be clustered in the confusion
matrix heatmap? Default is TRUE.}

\item{plot_cells}{logical, should cells be plotted as well? Default is FALSE
for speed.}
}
\value{
a ggplot with different types of visualization depending on the
    classes of the columns chosen. If y is a numeric and x is NULL, or a
    categorical (character/facrtor), the function plots a violin + boxplot divided by
    x and/or by color_by. If y and x are both numeric, the function plots a
    scatterplot with optional 2D kernel contouring. If y and x are both
    categorical, the function plots a heatmap of the confusion matrix,
    showing pairwise Jaccard indices. All plots can be facetted.
}
\description{
Plot data from the colData of the SCE object
}
