% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{Plot UMAP}
\usage{
plot_UMAP(
  sce,
  umap_slot = "UMAP",
  color_by = NULL,
  shape_by = NULL,
  group_by = NULL,
  label_by = NULL,
  color_palette = NULL
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{umap_slot}{character, the name in the reducedDim slot of the SCE object, e.g.
"UMAP_Harmony"}

\item{color_by}{character, column name in the colData slot of the SCE object, 
e.g. "cluster". Will be used to assign colors. Automatically detects 
whether the variable is categorical or continuous.}

\item{shape_by}{character, column name in the colData slot of the SCE object,
e.g. "cluster". Will be used to assign shapes. Can only be used for 
categorical variables.}

\item{group_by}{character, column name in the colData slot of the SCE object,
e.g. "individual". Will be used to facet the plot. Can only be used for
categorical variables.}

\item{label_by}{character, column name in the colData slot of the SCE object,
e.g. "cluster". Will be used to add labels to the plot. Can only be used
for categorical variables.}

\item{color_palette}{a character string containing colors to be used. Default
is NULL, meaning an automatic palette will be generated.}
}
\value{
a ggplot object showing the UMAP coordinates colored, shaped and 
    faceted as set by the arguments.
}
\description{
Plots the UMAP from a SingleCellExperiment object
}
