% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{Plot UMAP}
\usage{
plot_UMAP(
  sce,
  umap_slot = "UMAP",
  color_by = NULL,
  shape_by = NULL,
  group_by = NULL,
  label_by = NULL,
  point_size = 0.7,
  label_size = 2,
  color_palette = NULL,
  trajectory = NULL,
  rescale = TRUE
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{umap_slot}{character, the name in the reducedDim slot of \code{sce}, e.g.
\code{"UMAP_Harmony"}}

\item{color_by}{character, column name in the \code{colData} slot of \code{sce},
e.g. "cluster". Will be used to assign colors. Automatically detects
whether the variable is categorical or continuous.}

\item{shape_by}{character, column name in the \code{colData} slot of  \code{sce},
e.g. "cluster". Will be used to assign shapes. Can only be used for
categorical variables.}

\item{group_by}{character, column name in the \code{colData} slot of the \code{sce},
e.g. "individual". Will be used to facet the plot. Can only be used for
categorical variables.}

\item{label_by}{character, column name in the \code{colData} slot of \code{sce},
e.g. "cluster". Will be used to add labels to the plot. Can only be used
for categorical variables.}

\item{point_size}{numeric, the size of the points in the plot. Default is 0.7.}

\item{label_size}{numeric, the size of the font for the labels. Default is 2.}

\item{color_palette}{a character string containing colors to be used. Default
is \code{NULL}, meaning an automatic palette will be generated.}

\item{trajectory}{a character string indicating the `metadata` slot containing
segment trajectories to be plotted. Usually either 
\code{"Slingshot_embedded_curves"} or \code{"Monocle_embedded_curves"}.}

\item{rescale}{logical, should coordinates be rescaled between 0 and 1? Default is TRUE.}
}
\value{
a ggplot object showing the UMAP coordinates colored, shaped and
    faceted as set by the arguments.
}
\description{
Plots the UMAP from a SingleCellExperiment object
}
