% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotModularity}
\alias{plotModularity}
\title{Plot cluster modularity}
\usage{
plotModularity(sce, name, type = "heatmap")
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{name}{character, the name in the \code{metadata slot} of the
\code{SingleCellExperiment} object, e.g. "modularity_SNN_100"}

\item{type}{character, one of \code{"heatmap"} or \code{"graph"}
for different plot types}
}
\value{
a heatmap of pairwise modularity
}
\description{
Plots a heatmap showing pairwise cluster modularity
}
